requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.0
pandas==2.2.3
gunicorn==23.0.0
whitenoise==6.9.0